criando relacionamentos e repositories para as entidades

+++ src/main/java/com/projectspring/aula/domain/entity/ItemPedido.java
@@ -8,7 +8,7 @@ public class ItemPedido {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private long id;
+    private Integer id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "pedido_id")
@@ -21,11 +21,11 @@ public class ItemPedido {
     @Column
     private Integer quantity;
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
